+++ mapboard/topology_manager/procedures/reset-linework-errors.sql
@@ -1,1 +1,4 @@
-UPDATE {data_schema}.linework SET topology_error = null
+UPDATE {boundary_table} l
+SET topology_error = null
+WHERE l.topology_error IS NOT null
+  AND ({row_filter})